Move SharedCompilationServerOptions methods to a separate object
