+++ modules/cli/src/main/scala/scala/cli/commands/BloopStartOptions.scala
@@ -6,6 +6,7 @@ import scala.build.blooprifle.BloopRifleConfig
 import scala.build.options.{BuildOptions, InternalOptions}
 import scala.cli.commands.util.CommonOps._
 import scala.cli.commands.util.JvmUtils
+import scala.cli.commands.util.SharedCompilationServerOptionsUtil._
 
 // format: off
 final case class BloopStartOptions(
