+++ modules/cli/src/main/scala/scala/cli/commands/util/SharedOptionsUtil.scala
@@ -17,6 +17,7 @@ import scala.build.options.{Platform, ScalacOpt, ShadowingSeq}
 import scala.build.{options => bo}
 import scala.cli.commands.ScalaJsOptions
 import scala.cli.commands.util.CommonOps._
+import scala.cli.commands.util.SharedCompilationServerOptionsUtil._
 import scala.concurrent.duration._
 import scala.util.Properties
 import scala.util.control.NonFatal
